backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
pytesseract==0.3.10
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
